fix map name generation
