+++ pkg/security/secl/model/consts_linux.go
@@ -1777,84 +1777,6 @@ const (
 	BpfSkSkbVerdict
 )
 
-var bpfMapNames = []string{
-	"activity_dump_c",
-	"activity_dump_r",
-	"activity_dumps_",
-	"basename_approv",
-	"bb_approver_sta",
-	"bb_discarder_st",
-	"bpf_maps",
-	"bpf_progs",
-	"buffer_selector",
-	"cgroup_tracing_",
-	"cgroup_wait_lis",
-	"conntrack",
-	"discarders_revi",
-	"dns_event",
-	"dr_erpc_buffer",
-	"dr_erpc_state",
-	"dr_erpc_stats_b",
-	"dr_erpc_stats_f",
-	"enabled_events",
-	"events",
-	"events_ringbuf_",
-	"events_stats",
-	"exec_file_cache",
-	"exec_pid_transf",
-	"fb_approver_sta",
-	"fb_discarder_st",
-	"fd_link_pid",
-	"filter_policy",
-	"flow_pid",
-	"inode_disc_revi",
-	"inode_discarder",
-	"io_uring_ctx_pi",
-	"is_discarded_by",
-	"is_new_kthread",
-	"kill_list",
-	"mmap_flags_appr",
-	"mmap_protection",
-	"mount_ref",
-	"mprotect_req_pr",
-	"mprotect_vm_pro",
-	"netdevice_looku",
-	"netns_cache",
-	"open_flags_appr",
-	"packets",
-	"path_id",
-	"pathnames",
-	"pid_cache",
-	"pid_discarders",
-	"pid_ignored",
-	"proc_cache",
-	"process_event_g",
-	"register_netdev",
-	"secprofs_syscal",
-	"security_profil",
-	"selinux_enforce",
-	"selinux_write_b",
-	"span_tls",
-	"splice_entry_fl",
-	"splice_exit_fla",
-	"str_array_buffe",
-	"syscall_monitor",
-	"syscall_table",
-	"syscalls",
-	"syscalls_stats",
-	"syscalls_stats_",
-	"tasks_in_coredu",
-	"tgid_fd_map_id",
-	"tgid_fd_prog_id",
-	"traced_cgroups",
-	"traced_comms",
-	"traced_pids",
-	"user_sessions",
-	"veth_device_nam",
-	"veth_devices",
-	"veth_state_mach",
-}
-
 var capsStringArrayCache *lru.Cache[KernelCapability, []string]
 
 func init() {
